# refactor: Preallocate addresses in GetAddr based on nNodes