@paplorinc 2024-03-09
@brunoerg, feel free to do that, I'll close this if you cherry-pick it to the other PR (please provide before/after measurements)