@fanquake 2024-10-25
> I don't think this is a particularly impactful change but it's very small and straightforward to review.

I agree, and I think we are probably overdue for a review of what is in `src/bench`, and cleaning out benchmarks that aren't (very) meaningful. IIRC `clang-tidy` has a check for exactly this kind of change, so I think if any more were going to be made, they could be done using that.